requirements.txt
Django==5.1.1
python-dotenv==1.0.1
psycopg2==2.9.9
django-unfold==0.39.0